+++ guillotina/contrib/redis/driver.py
@@ -1,11 +1,11 @@
 try:
-    import aioredis
+    import redis.asyncio as aioredis
 except ImportError:
-    print("If you add guillotina.contrib.redis you need to add aioredis on your requirements")
+    print("If you add guillotina.contrib.redis you need to add redis>4.2.0rc1 on your requirements")
     raise
 
-from aioredis.client import PubSub
-from aioredis.exceptions import ConnectionError
+from redis.client import PubSub
+from redis.exceptions import ConnectionError
 from guillotina import app_settings
 from guillotina import metrics
 from guillotina.contrib.redis.exceptions import NoRedisConfigured